requirements.txt
django>=5.2.1
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
pillow>=11.2.1
reportlab>=4.4.0
python-dotenv>=1.0.0
six>=1.16.0
django-rest-auth>=0.9.5
djangorestframework-simplejwt>=5.3.1